requirements.txt
pyspark>=3.0.0
dbt-core>=1.0.0
dbt-duckdb>=1.0.0
pandas>=1.3.0
pyarrow>=5.0.0
apache-airflow>=2.2.0
